+++ app/Http/Controllers/Admin/InsuranceController.php
@@ -8,35 +8,38 @@ use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Models\InsurancePost;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 class InsuranceController extends Controller
 {
     public function index()
     {
+        Gate::authorize('app.insurances.index');
         $insurances = Insurance::all();
         return view('admin.insurance.index', compact('insurances'));
     }
 
     public function create()
     {
+        Gate::authorize('app.insurances.create');
         return view('admin.insurance.create');
     }
 
     public function store(Request $request)
     {
-        // return $request->all();
+        Gate::authorize('app.insurances.create');
         $request->validate([
             'title' => 'required|max:255',
             'status' => 'required',
             'header_image' => 'mimes:jpeg,png,jpg,gif'
         ]);
-        
+
         $header_image = $request->file('header_image');
         if (isset($header_image)) {
             $currentDate = Carbon::now()->toDateString();
                 $headerimagename = $currentDate.uniqid().'.'.$header_image->getClientOriginalExtension();
-    
+
                 if (!Storage::disk('public')->exists('frontend/assets/img/insurances')) {
                     Storage::disk('public')->makeDirectory('frontend/assets/img/insurances');
                 }
@@ -61,28 +64,30 @@ class InsuranceController extends Controller
         }
 
         return redirect()->route('admin.insurance.index');
-        
+
     }
 
     public function edit($id)
     {
+        Gate::authorize('app.insurances.edit');
         $insurance = Insurance::findOrFail($id);
         return view('admin.insurance.edit',compact('insurance'));
     }
 
     public function update(Request $request, $id)
     {
+        Gate::authorize('app.insurances.edit');
         $insurance = Insurance::findOrfail($id);
         $header_image = $request->file('header_image');
         $old_image = $insurance->header_image;
         if (isset($header_image)) {
             $currentDate = Carbon::now()->toDateString();
                 $headerimagename = $currentDate.uniqid().'.'.$header_image->getClientOriginalExtension();
-    
+
                 if (!Storage::disk('public')->exists('frontend/assets/img/insurances')) {
                     Storage::disk('public')->makeDirectory('frontend/assets/img/insurances');
                 }
-                
+
                 if (Storage::disk('public')->exists('frontend/assets/img/insurances/'.$old_image)) {
                     Storage::disk('public')->delete('frontend/assets/img/insurances/'.$old_image);
                 }
@@ -90,7 +95,7 @@ class InsuranceController extends Controller
                 Storage::disk('public')->putFileAs('frontend/assets/img/insurances/',$header_image,$headerimagename);
 
                 $insurance->header_image = $headerimagename;
-                
+
                 $data = [
                     'title' => $request->title,
                     'url' => Str::slug($request->title),
@@ -113,26 +118,29 @@ class InsuranceController extends Controller
 
     public function destroy($id)
     {
+        Gate::authorize('app.insurances.destroy');
         $insurance = Insurance::findOrFail($id)->delete();
         return redirect()->back();
     }
 
-    // Insurance posts 
+    // Insurance posts
     public function InsurancePostIndex()
     {
+        Gate::authorize('app.insurances.posts.index');
         $insurancePosts = InsurancePost::all();
         return view('admin.insurance.posts.index',compact('insurancePosts'));
     }
 
     public function InsurancePostCreate()
     {
+        Gate::authorize('app.insurances.posts.create');
         $insurances = Insurance::all();
         return view('admin.insurance.posts.create',compact('insurances'));
     }
 
     public function InsurancePostStore(Request $request)
     {
-        // return $request->all();
+        Gate::authorize('app.insurances.posts.create');
 
         $request->validate([
             'title' => 'required|max:255',
@@ -178,6 +186,7 @@ class InsuranceController extends Controller
 
     public function InsurancePostEdit($id)
     {
+        Gate::authorize('app.insurances.posts.edit');
         $insurances = Insurance::all();
         $insurancePost = InsurancePost::findOrFail($id);
         return view('admin.insurance.posts.edit',compact('insurancePost','insurances'));
@@ -185,6 +194,7 @@ class InsuranceController extends Controller
 
     public function InsurancePostUpdate(Request $request, $id)
     {
+        Gate::authorize('app.insurances.posts.edit');
         $insurance = InsurancePost::findOrFail($id);
         $featured_image = $request->file('featured_image');
         $pdf_file = $request->file('pdf_file');
@@ -196,7 +206,7 @@ class InsuranceController extends Controller
             if (!Storage::disk('public')->exists('frontend/assets/img/insurances/posts')) {
                 Storage::disk('public')->makeDirectory('frontend/assets/img/insurances/posts');
             }
-            
+
             if (Storage::disk('public')->exists('frontend/assets/img/insurances/posts/'.$old_image)) {
                 Storage::disk('public')->delete('frontend/assets/img/insurances/posts/'.$old_image);
             }
@@ -234,6 +244,7 @@ class InsuranceController extends Controller
 
     public function InsurancePostDestroy($id)
     {
+        Gate::authorize('app.insurances.posts.destroy');
         $insurance = InsurancePost::findOrFail($id);
         $old_image = $insurance->featured_image;
         $old_file = $insurance->pdf_file;
